Info Heirarchy Changes & Expansion
Moved the various info models into the cluster_info package.
Added some missing fields to the table & keyspace models.
